cleanup, fix filtering menu
