+++ src/lib/fns.ts
@@ -1,0 +1,71 @@
+import { POOLS } from 'constants/farms'
+import {ChainId} from 'sdk'
+
+export const farmMap = (summonerInfo: any, soulPrice: number, chainId: ChainId) => (pool: any) => {
+  pool.owner = 'SoulSwap'
+  pool.balance = 0
+
+  const pair = POOLS[chainId][pool.lpToken]
+
+//   const secondsPerHour = 60 * 60
+
+  function getRewards() {
+    const rewardPerSecond =
+      ((pool.allocPoint / Number(summonerInfo.totalAllocPoint)) * Number(summonerInfo.soulPerSecond)) / 1e18
+
+    const defaultReward = {
+      token: 'SOUL',
+      icon: '/images/token/soul.png',
+      rewardPerSecond,
+      rewardPerDay: rewardPerSecond * 86400,
+      rewardPrice: soulPrice,
+    }
+
+    const defaultRewards = [defaultReward]
+
+    return defaultRewards
+  }
+
+  const rewards = getRewards()
+
+//   // const tvl = getTvl()
+//   const tvl = pool.pair?.token1
+//   ? Number(pool.pairPrice) * Number(pool.lpBalance) / 1e18
+//   : Number(soulPrice) * Number(pool.lpBalance) / 1e18
+
+//   const rewardPerSec =
+//     ((pool.allocPoint / Number(summonerInfo.totalAllocPoint)) * Number(summonerInfo.soulPerSecond)) / 1e18
+
+//   const rewardPrice = soulPrice
+
+//   const roiPerSecond =
+//     farms.reduce((previousValue, currentValue) => {
+//       return previousValue + rewardPerSec * rewardPrice
+//     }, 0) / Number(tvl)
+
+//   // const roiPerSecond = Number(tvl)
+//   const roiPerHour = roiPerSecond * 60 * 60
+//   const roiPerDay = roiPerHour * 24
+//   const roiPerMonth = roiPerDay * 30
+//   const roiPerYear = roiPerDay * 365
+
+//   const position = positions.find((position) => position.id === pool.id)
+
+  return {
+    ...pool,
+    // ...position,
+    allocPointNumber: pool.allocPoint?.toNumber ? pool.allocPoint.toNumber() : pool.allocPoint || 0,
+    pair: {
+      ...pair,
+      decimals: 18,
+    },
+    // roiPerSecond,
+    // roiPerHour,
+    // roiPerDay,
+    // roiPerMonth,
+    // roiPerYear,
+    rewards,
+    // tvl,
+    // secondsPerHour,
+  }
+}
